Optimize lock judge engine.
